Store information about login so that it can be reused
