+++ .gitignore
@@ -11,3 +11,4 @@ node_modules/
 
 .env
 playwright-screenshots
+e2e/.auth.json
