+++ playwright.config.ts
@@ -7,7 +7,7 @@ import { defineConfig, devices } from '@playwright/test';
  */
 require('dotenv').config();
 
-export const STORAGE_STATE = pjoin(__dirname, '/user.json');
+export const STORAGE_STATE = pjoin(__dirname, 'e2e', '.auth.json');
 
 /**
  * See https://playwright.dev/docs/test-configuration.
